Dump jasmine gem in favor of a runner in the app
- Uses application.js for JS dependencies
- New specs go in app/assets/specs/**
- Visit /specs to run specs in browser

I couldn't figure out how to get the jasmine gem to
play nicely with the asset pipeline, and I didn't
want to specify dependencies in multiple places. I
also didn't enjoy having to start a separate server
for running specs.

I still want to extract these pieces out so they are
not intertwined with production code, but for now
this works.